+++ MultiplicationTree.h
@@ -19,7 +19,7 @@ class MultiplicationTree
         string description;
         int delay(shared_ptr<Node>);
         string expression(shared_ptr<Node>);
-        long long execute(shared_ptr<Node>, long long, long long);
+        __int128 execute(shared_ptr<Node>, __int128, __int128);
 
     public:
         MultiplicationTree();
@@ -30,7 +30,7 @@ class MultiplicationTree
         string getExpression();
         void setRoot(shared_ptr<OperationNode>);
 	void setDescription(string);	
-        long long executeMultiplication(long long, long long);
+        __int128 executeMultiplication(__int128, __int128);
         virtual ~MultiplicationTree() = default;
 };
